@@ -11,17 +11,17 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^8.1.3",
+    "@angular/animations": "^8.2.3",
     "@angular/cdk": "^8.1.2",
-    "@angular/common": "~8.1.3",
-    "@angular/compiler": "~8.1.3",
-    "@angular/core": "~8.1.3",
+    "@angular/common": "~8.2.3",
+    "@angular/compiler": "~8.2.3",
+    "@angular/core": "~8.2.3",
     "@angular/fire": "^5.2.1",
-    "@angular/forms": "~8.1.3",
+    "@angular/forms": "~8.2.3",
     "@angular/material": "^8.1.2",
-    "@angular/platform-browser": "~8.1.3",
-    "@angular/platform-browser-dynamic": "~8.1.3",
-    "@angular/router": "~8.1.3",
+    "@angular/platform-browser": "~8.2.3",
+    "@angular/platform-browser-dynamic": "~8.2.3",
+    "@angular/router": "~8.2.3",
     "@ngrx/effects": "^8.1.0",
     "@ngrx/store": "^8.1.0",
     "@ngrx/store-devtools": "^8.1.0",
@@ -35,10 +35,10 @@
     "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.801.3",
-    "@angular/cli": "~8.1.3",
-    "@angular/compiler-cli": "~8.1.3",
-    "@angular/language-service": "~8.1.3",
+    "@angular-devkit/build-angular": "~0.803.0",
+    "@angular/cli": "~8.3.0",
+    "@angular/compiler-cli": "~8.2.3",
+    "@angular/language-service": "~8.2.3",
     "@types/node": "~8.9.4",
     "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.3",
@@ -53,6 +53,6 @@
     "protractor": "~5.4.0",
     "ts-node": "~7.0.0",
     "tslint": "~5.11.0",
-    "typescript": "~3.4.5"
+    "typescript": "~3.5.3"
   }
 }
