@@ -1,7 +1,7 @@
 import { Injectable } from '@angular/core';
 import { Store } from '@ngrx/store';
 
-import { Bug } from '../Bug';
+import { Bug, UpdateFilteredList } from '../Bug';
 import { GetBugs } from '../Bug';
 
 interface BugsState {
@@ -16,5 +16,9 @@ export class BugsService {
 
   getBugs() {
     this.store.dispatch(new GetBugs());
+  }
+
+  updateFilteredList(filters: string[]) {
+    this.store.dispatch(new UpdateFilteredList(filters));
   }
 }
